+++ examples/Bancontact/Controller.php
@@ -1,9 +1,9 @@
 <?php
 /**
- * This is the controller for the Invoice guaranteed example.
+ * This is the controller for the Bancontact example.
  * It is called when the pay button on the index page is clicked.
  *
- * Copyright (C) 2019 heidelpay GmbH
+ * Copyright (C) 2020 heidelpay GmbH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,7 +19,7 @@
  *
  * @link  https://docs.heidelpay.com/
  *
- * @author  Simon Gabriel <development@heidelpay.com>
+ * @author  David Owusu <development@heidelpay.com>
  *
  * @package  heidelpayPHP\examples
  */
@@ -34,8 +34,6 @@ require_once __DIR__ . '/../../../../autoload.php';
 use heidelpayPHP\examples\ExampleDebugHandler;
 use heidelpayPHP\Exceptions\HeidelpayApiException;
 use heidelpayPHP\Heidelpay;
-use heidelpayPHP\Resources\Basket;
-use heidelpayPHP\Resources\EmbeddedResources\BasketItem;
 
 session_start();
 session_unset();
@@ -52,11 +50,10 @@ function redirect($url, $merchantMessage = '', $clientMessage = '')
 }
 
 // You will need the id of the payment type created in the frontend (index.php)
-if (!isset($_POST['paymentTypeId'], $_POST['customerId'])) {
+if (!isset($_POST['resourceId'])) {
     redirect(FAILURE_URL, 'Resource id is missing!', $clientMessage);
 }
-$paymentTypeId   = $_POST['paymentTypeId'];
-$customerId  = $_POST['customerId'];
+$paymentTypeId   = $_POST['resourceId'];
 
 // Catch API errors, write the message to your log and show the ClientMessage to the client.
 try {
@@ -64,44 +61,24 @@ try {
     $heidelpay = new Heidelpay(HEIDELPAY_PHP_PAYMENT_API_PRIVATE_KEY);
     $heidelpay->setDebugMode(true)->setDebugHandler(new ExampleDebugHandler());
 
-    $orderId = 'o' . str_replace(['0.', ' '], '', microtime(false));
+    // Create a charge transaction to get the redirectUrl.
+    $transaction = $heidelpay->charge(12.32, 'EUR', $paymentTypeId, RETURN_CONTROLLER_URL);
 
-    // A Basket is mandatory for Invoice Factoring payment type
-    $basketItem = (new BasketItem('Hat', 100.00, 119.00, 1))
-        ->setAmountGross(119.0)
-        ->setAmountVat(19.0);
-    $basket = new Basket($orderId, 119.0, 'EUR', [$basketItem]);
-
-    $transaction = $heidelpay->charge(119.0, 'EUR', $paymentTypeId, CONTROLLER_URL, $customerId, $orderId, null, $basket);
-
-    // You'll need to remember the shortId to show it on the success or failure page
-    $_SESSION['ShortId'] = $transaction->getShortId();
+    // You'll need to remember the paymentId for later in the ReturnController
     $_SESSION['PaymentId'] = $transaction->getPaymentId();
-    $_SESSION['additionalPaymentInformation'] =
-        sprintf(
-            "Please transfer the amount of %f %s to the following account:<br /><br />"
-            . "Holder: %s<br/>"
-            . "IBAN: %s<br/>"
-            . "BIC: %s<br/><br/>"
-            . "<i>Please use only this identification number as the descriptor: </i><br/>"
-            . "%s",
-            $transaction->getAmount(),
-            $transaction->getCurrency(),
-            $transaction->getHolder(),
-            $transaction->getIban(),
-            $transaction->getBic(),
-            $transaction->getDescriptor()
-        );
+    $_SESSION['ShortId']   = $transaction->getShortId();
 
-    // To avoid redundant code this example redirects to the general ReturnController which contains the code example to handle payment results.
-    redirect(RETURN_CONTROLLER_URL);
+    // Redirect to the Bancontact page
+    if (!$transaction->isError() && $transaction->getRedirectUrl() !== null) {
+        redirect($transaction->getRedirectUrl());
+    }
 
+    // Check the result message of the charge to find out what went wrong.
+    $merchantMessage = $transaction->getMessage()->getCustomer();
 } catch (HeidelpayApiException $e) {
     $merchantMessage = $e->getMerchantMessage();
     $clientMessage = $e->getClientMessage();
 } catch (RuntimeException $e) {
     $merchantMessage = $e->getMessage();
 }
-// Write the merchant message to your log.
-// Show the client message to the customer (it is localized).
 redirect(FAILURE_URL, $merchantMessage, $clientMessage);