+++ examples/HirePurchaseDirectDebit/Controller.php
@@ -87,8 +87,7 @@ try {
         ->setEmail('linda.heideich@test.de');
 
     // A Basket is mandatory for SEPA direct debit guaranteed payment type
-    $basketItem = (new BasketItem('Hat', 100.00, 100.00, 1))
-        ->setAmountNet(100.0)
+    $basketItem = (new BasketItem('Hat', 100.00, 119.00, 1))
         ->setAmountGross(119.0)
         ->setAmountVat(19.0);
     $basket = (new Basket($orderId, 119.0, 'EUR', [$basketItem]))->setAmountTotalVat(19.0);