+++ examples/HostedPayPage/Controller.php
@@ -88,7 +88,9 @@ try {
     $paypage->setEffectiveInterestRate(4.99);
 
     // ... a Basket is mandatory for HirePurchase
-    $basketItem = new BasketItem('Hat', 100.0, 119.0, 1);
+    $basketItem = (new BasketItem('Hat', 100.00, 119.00, 1))
+        ->setAmountGross(119.0)
+        ->setAmountVat(19.0);
     $basket = new Basket($orderId, 119.0, 'EUR', [$basketItem]);
 
     if ($transactionType === 'charge') {