+++ src/Constants/IdStrings.php
@@ -52,6 +52,7 @@ class IdStrings
     public const INVOICE_FACTORING = 'ivf';
     public const HIRE_PURCHASE_DIRECT_DEBIT = 'hdd';
     public const PAYMENT_PAGE = 'ppg';
+    public const BANCONTACT = 'bct';
 
     // Resources
     public const BASKET = 'bsk';
@@ -78,6 +79,7 @@ class IdStrings
         self::WECHATPAY,
         self::INVOICE_FACTORING,
         self::PAYMENT_PAGE,
-        self::HIRE_PURCHASE_DIRECT_DEBIT
+        self::HIRE_PURCHASE_DIRECT_DEBIT,
+        self::BANCONTACT
     ];
 }