+++ src/Heidelpay.php
@@ -63,7 +63,7 @@ class Heidelpay implements HeidelpayParentInterface, PaymentServiceInterface, Re
     public const BASE_URL = 'api.heidelpay.com';
     public const API_VERSION = 'v1';
     public const SDK_TYPE = 'heidelpayPHP';
-    public const SDK_VERSION = '1.2.8.0';
+    public const SDK_VERSION = '1.2.9.0';
 
     /** @var string $key */
     private $key;