+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.2.9.0][1.2.9.0]
+
+### Add
+*   Bancontact payment type and example.
+
+### Fix
+*   Payment page examples.
+
 ## [1.2.8.0][1.2.8.0]
 
 ### Add
@@ -444,3 +452,4 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 [1.2.7.1]: https://github.com/heidelpay/heidelpayPHP/compare/1.2.7.0..1.2.7.1
 [1.2.7.2]: https://github.com/heidelpay/heidelpayPHP/compare/1.2.7.1..1.2.7.2
 [1.2.8.0]: https://github.com/heidelpay/heidelpayPHP/compare/1.2.7.2..1.2.8.0
+[1.2.9.0]: https://github.com/heidelpay/heidelpayPHP/compare/1.2.8.0..1.2.9.0